% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eia_oil_prices}
\alias{eia_oil_prices}
\title{EIA Oil Spot Prices}
\format{
A list of time series.
}
\source{
\href{https://www.eia.gov/dnav/pet/pet_pri_spt_s1_d.htm}{EIA Energy Information Administration}
}
\usage{
data(eia_oil_prices)
}
\description{
Spot prices of oil and other liquids such as gasoline, heating oil, diesel, jet fuel and propane.
Crude oil in Dollars per barrel, other products in Dollars per gallon. Prices are calculated
by EIA from daily data by taking an unweighted average of the daily closing spot prices
for a given product over the specified time period.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? No
}
\examples{
data(eia_oil_prices)
serie <- eia_oil_prices[[1]]
}
\keyword{#'}
\keyword{@references}
\keyword{Administration}
\keyword{EIA}
\keyword{Energy}
\keyword{Information}
\keyword{datasets}
