% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fi_br}
\alias{fi_br}
\title{Finance Indexes}
\format{
A list of time series.
}
\source{
\href{https://github.com/cefet-rj-dal/event_datasets/raw/main/original_datasets/fi_br/data/glossary.xlsx}{Glossary with complete references}
}
\usage{
data(fi_br)
}
\description{
Finance: Securities Price Indexes (Brazil).
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(fi_br)
serie <- fi_br[[1]]
}
\keyword{#'#'}
\keyword{(b3,}
\keyword{@references}
\keyword{Bank}
\keyword{Br}
\keyword{Brazil)}
\keyword{Central}
\keyword{Investing,}
\keyword{Various}
\keyword{datasets}
\keyword{of}
\keyword{public}
\keyword{sources}
