% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gecco}
\alias{gecco}
\title{Gecco Challenge 2018}
\format{
A list of time series.
}
\source{
\href{https://www.spotseven.de/gecco/gecco-challenge}{Gecco Challenge 2018}
}
\usage{
data(gecco)
}
\description{
Data collection for water quality monitoring.
Data Type:	Water quality. Category:	Environment. Creation Date	2018.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(gecco)
serie <- gecco[[1]]
}
\keyword{#'}
\keyword{(ACM)}
\keyword{(Gecco),}
\keyword{@references}
\keyword{Association}
\keyword{Computation}
\keyword{Computing}
\keyword{Conference}
\keyword{Evolutionary}
\keyword{Genetic}
\keyword{Machinery}
\keyword{and}
\keyword{datasets}
\keyword{for}
