% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality_cnes}
\alias{mortality_cnes}
\title{Child mortality dataset}
\format{
A list of time series.
}
\usage{
data(mortality_cnes)
}
\description{
Child mortality dataset. Data for assessing infant mortality,
an important health indicator.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(mortality_cnes)
serie <- mortality_cnes[[1]]
}
\keyword{datasets}
