% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oil_3w_Type_1}
\alias{oil_3w_Type_1}
\title{Oil wells dataset - Type 1}
\format{
A list of time series.
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/3W+dataset}{UCI Machine Learning Repository}
}
\usage{
data(oil_3w_Type_1)
}
\description{
Firts realistic dataset with real events in oil well drilling. The data available
in this package consist of time series already analyzed and applied in
research experiments by the DAL group (Data Analytics Lab). The series are
divided into 7 groups (Type_0, Type_1, Type_2, Type_5, Type_6, Type_7 and Type_8).
Creation date: 2019.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(oil_3w_Type_1)
serie <- oil_3w_Type_1[[1]]
}
\keyword{#'}
\keyword{3W}
\keyword{@references}
\keyword{Data}
\keyword{Set}
\keyword{dataset}
\keyword{datasets}
