% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rare}
\alias{rare}
\title{RARE: A Labeled Dataset for Cloud-Native Memory Anomalies}
\format{
A list of time series.
}
\source{
\href{https://doi.org/10.1145/3416505.3423560}{RARE: A Labeled Dataset for Cloud-Native Memory Anomalies}
}
\usage{
data(rare)
}
\description{
Data with over 900 metrics generated injecting artificial anomalies
in real streaming services. Data is organized into over 7000 variables.
For a detailed list of variables see:
\href{https://github.com/cefet-rj-dal/event_datasets_etl/blob/main/original_datasets/rare/data/rare_attributes.csv}{RARE Attributes}.
Real data from stream services and artificial labeled anomalies.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(rare)
serie <- rare$kafka_network_processor_idlepercent_value_0
}
\keyword{#'#'}
\keyword{@references}
\keyword{RARE}
\keyword{datasets}
