% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucr}
\alias{ucr}
\title{UCR Anomaly Archive}
\format{
A list of time series.
}
\source{
\href{https://paperswithcode.com/dataset/ucr-anomaly-archive}{UCR Anomaly Archive}
}
\usage{
data(ucr)
}
\description{
Data with 250 series with real world data and synthetic and real anomalies.
Real data from human medicine, biology, meteorology and industry.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(ucr)
serie <- ucr[[1]]
}
\keyword{#'#'}
\keyword{@references}
\keyword{Anomaly}
\keyword{Archive}
\keyword{UCR}
\keyword{datasets}
