% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yahoo_a1}
\alias{yahoo_a1}
\title{A1 Yahoo Benchmark Dataset for Time Series Anomaly Detection}
\format{
A list of time series.
}
\source{
\href{https://yahooresearch.tumblr.com/post/114590420346/a-benchmark-dataset-for-time-series-anomaly}{Yahoo Labs}
}
\usage{
data(yahoo_a1)
}
\description{
A1 Group --> Real data from Yahoo services internet traffic.
Data collected from user activity and server logs from Yahoo services.
Internet traffic and synthetic data. Creation Date	2015.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(yahoo_a1)
serie <- yahoo_a1[[1]]
}
\references{
Yahoo Webscope data-sharing program
}
\keyword{datasets}
